% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCNmodel.R
\name{PCNmodel}
\alias{PCNmodel}
\title{PCNmodel}
\usage{
PCNmodel(
  numYears,
  simStartDay,
  initialEggsPerGram,
  soil.density = 1.5,
  plantingDOY = 90,
  harvestingDOY = 216,
  noPotatoYearsList = list(c(start = 2, fin = 3)),
  lifecycle.pars = list(cyst.pause = 365, eggs_per_cyst = 250, deathC = 5e-04, Copt =
    0.88, gen2 = 0, tauA.m = 21, tauA.f = 21, PA = 1, dormancy = 5e-04),
  ode.compartments = list(kE = 20, kJ = 20, kA = 50),
  temperature.pars = list(temperaturefile = NULL, fixTemp = 14, deltaT = 0),
  plant.pars = list(resistanceFactor = 0, N.plants = 1, root.length = 500, roots.per.cm
    = 0.5)
)
}
\arguments{
\item{numYears}{number of years to simulate}

\item{simStartDay}{day of the year the simulation starts. Defaults to 0}

\item{initialEggsPerGram}{Initial egg density. Defaults to 0.1 UNITS???}

\item{soil.density}{Soil density (UNITS??) Defaults to 1.5}

\item{plantingDOY}{90, # potatoes planted on this day of the year}

\item{harvestingDOY}{216, # potatoes harvested on this day of the year}

\item{noPotatoYearsList}{Defaults to list(c(start=2,fin=3)), # Choose years not to plant potatoes. Can compare the different strategies by adding more to this list. For continuous planting use c(start=0,fin=0).}

\item{lifecycle.pars}{list containing parameters controlling the PCN lifecycle
\itemize{
\item cyst.pause = 365 length of obligatory diapause (days)
\item eggs_per_cyst = 250  number of eggs per cyst
\item deathC = 0.0005 mortality rate of cysts
\item Copt = 0.88 proportion of cysts which hatch under optimum conditions
\item gen2 = 0 proportion of cysts without a diapause requirement
\item tauA.m = 21 Duration of adult male stage
\item tauA.f = 21 Duration of adult female stage
\item PA = 1 Proportion of adults which survive
\item dormancy = 0.0005 cyst activation from dormancy
}}

\item{ode.compartments}{list of how many ODEs are needed for the different life stages
\itemize{
\item kE = 20, number of compartments in egg stage
\item kJ = 20, number of compartments in juvenile stage
\item kA = 50, number of compartments in adult stage
}}

\item{temperature.pars}{list}

\item{plant.pars}{list of factors relating to the potato plant
\itemize{
\item resistanceFactor = 0, resistance factor of potatoes
\item N.plants = 1, Number of potato plants
\item root.length = 500, maximum length of plant roots (cm) in healthy plant
\item oots.per.cm = 0.5, root length density of potato plants
}}
}
\value{
matrix
}
\description{
PCNmodel
}
